Use separate data directory for MP3s
One mix can belong to multiple categories. Thus, a central data
directory is used, and the downloaded files are symlinked into
the style directories.

Blocking a certain mix can be done by putting an empty file in the
data directory, or in the style directory. In the latter case the
mix is only blocked for this style, and may be downloaded through
a different style.

Signed-off-by: Ralf Ertzinger <ralf@skytale.net>